refactor: moved utils used only in jsx_a11y to jsx_a11y.rs
