+++ crates/oxc_linter/src/utils/mod.rs
@@ -1,6 +1,7 @@
 mod jest;
+mod jsx_a11y;
 mod node;
 mod react;
 mod unicorn;
 
-pub use self::{jest::*, node::*, react::*, unicorn::*};
+pub use self::{jest::*, jsx_a11y::*, node::*, react::*, unicorn::*};
