+++ crates/oxc_linter/src/utils/react.rs
@@ -204,12 +204,3 @@ pub fn parse_jsx_value(value: &JSXAttributeValue) -> Result<f64, ()> {
         _ => Err(()),
     }
 }
-
-pub fn get_attribute_name(attr: &JSXAttributeName) -> String {
-    match attr {
-        JSXAttributeName::Identifier(ident) => ident.name.to_string(),
-        JSXAttributeName::NamespacedName(namespaced_name) => {
-            format!("{}:{}", namespaced_name.namespace.name, namespaced_name.property.name)
-        }
-    }
-}
